removes module name from direct
