+++ source/extensions/omni.isaac.lab_tasks/omni/isaac/lab_tasks/direct/ant/__init__.py
@@ -15,14 +15,12 @@ from . import agents
 # Register Gym environments.
 ##
 
-task_entry = "omni.isaac.lab_tasks.direct.ant"
-
 gym.register(
     id="Isaac-Ant-Direct-v0",
-    entry_point=f"{task_entry}.ant_env:AntEnv",
+    entry_point=f"{__name__}.ant_env:AntEnv",
     disable_env_checker=True,
     kwargs={
-        "env_cfg_entry_point": f"{task_entry}.ant_env:AntEnvCfg",
+        "env_cfg_entry_point": f"{__name__}.ant_env:AntEnvCfg",
         "rl_games_cfg_entry_point": f"{agents.__name__}:rl_games_ppo_cfg.yaml",
         "rsl_rl_cfg_entry_point": f"{agents.__name__}.rsl_rl_ppo_cfg:AntPPORunnerCfg",
         "skrl_cfg_entry_point": f"{agents.__name__}:skrl_ppo_cfg.yaml",
