+++ source/extensions/omni.isaac.lab_tasks/omni/isaac/lab_tasks/direct/cart_double_pendulum/__init__.py
@@ -15,14 +15,12 @@ from . import agents
 # Register Gym environments.
 ##
 
-task_entry = "omni.isaac.lab_tasks.direct.cart_double_pendulum"
-
 gym.register(
     id="Isaac-Cart-Double-Pendulum-Direct-v0",
-    entry_point=f"{task_entry}.cart_double_pendulum_env:CartDoublePendulumEnv",
+    entry_point=f"{__name__}.cart_double_pendulum_env:CartDoublePendulumEnv",
     disable_env_checker=True,
     kwargs={
-        "env_cfg_entry_point": f"{task_entry}.cart_double_pendulum_env:CartDoublePendulumEnvCfg",
+        "env_cfg_entry_point": f"{__name__}.cart_double_pendulum_env:CartDoublePendulumEnvCfg",
         "rl_games_cfg_entry_point": f"{agents.__name__}:rl_games_ppo_cfg.yaml",
         "skrl_cfg_entry_point": f"{agents.__name__}:skrl_ppo_cfg.yaml",
         "skrl_ippo_cfg_entry_point": f"{agents.__name__}:skrl_ippo_cfg.yaml",
