+++ source/extensions/omni.isaac.lab_tasks/omni/isaac/lab_tasks/direct/cartpole/__init__.py
@@ -15,14 +15,12 @@ from . import agents
 # Register Gym environments.
 ##
 
-task_entry = "omni.isaac.lab_tasks.direct.cartpole"
-
 gym.register(
     id="Isaac-Cartpole-Direct-v0",
-    entry_point=f"{task_entry}.cartpole_env:CartpoleEnv",
+    entry_point=f"{__name__}.cartpole_env:CartpoleEnv",
     disable_env_checker=True,
     kwargs={
-        "env_cfg_entry_point": f"{task_entry}.cartpole_env:CartpoleEnvCfg",
+        "env_cfg_entry_point": f"{__name__}.cartpole_env:CartpoleEnvCfg",
         "rl_games_cfg_entry_point": f"{agents.__name__}:rl_games_ppo_cfg.yaml",
         "rsl_rl_cfg_entry_point": f"{agents.__name__}.rsl_rl_ppo_cfg:CartpolePPORunnerCfg",
         "skrl_cfg_entry_point": f"{agents.__name__}:skrl_ppo_cfg.yaml",
@@ -32,10 +30,10 @@ gym.register(
 
 gym.register(
     id="Isaac-Cartpole-RGB-Camera-Direct-v0",
-    entry_point=f"{task_entry}.cartpole_camera_env:CartpoleCameraEnv",
+    entry_point=f"{__name__}.cartpole_camera_env:CartpoleCameraEnv",
     disable_env_checker=True,
     kwargs={
-        "env_cfg_entry_point": f"{task_entry}.cartpole_camera_env:CartpoleRGBCameraEnvCfg",
+        "env_cfg_entry_point": f"{__name__}.cartpole_camera_env:CartpoleRGBCameraEnvCfg",
         "rl_games_cfg_entry_point": f"{agents.__name__}:rl_games_camera_ppo_cfg.yaml",
         "skrl_cfg_entry_point": f"{agents.__name__}:skrl_camera_ppo_cfg.yaml",
     },
@@ -43,10 +41,10 @@ gym.register(
 
 gym.register(
     id="Isaac-Cartpole-Depth-Camera-Direct-v0",
-    entry_point=f"{task_entry}.cartpole_camera_env:CartpoleCameraEnv",
+    entry_point=f"{__name__}.cartpole_camera_env:CartpoleCameraEnv",
     disable_env_checker=True,
     kwargs={
-        "env_cfg_entry_point": f"{task_entry}.cartpole_camera_env:CartpoleDepthCameraEnvCfg",
+        "env_cfg_entry_point": f"{__name__}.cartpole_camera_env:CartpoleDepthCameraEnvCfg",
         "rl_games_cfg_entry_point": f"{agents.__name__}:rl_games_camera_ppo_cfg.yaml",
         "skrl_cfg_entry_point": f"{agents.__name__}:skrl_camera_ppo_cfg.yaml",
     },
