+++ source/extensions/omni.isaac.lab_tasks/omni/isaac/lab_tasks/direct/franka_cabinet/__init__.py
@@ -14,14 +14,13 @@ from . import agents
 # Register Gym environments.
 ##
 
-task_entry = "omni.isaac.lab_tasks.direct.franka_cabinet"
 
 gym.register(
     id="Isaac-Franka-Cabinet-Direct-v0",
-    entry_point=f"{task_entry}.franka_cabinet_env:FrankaCabinetEnv",
+    entry_point=f"{__name__}.franka_cabinet_env:FrankaCabinetEnv",
     disable_env_checker=True,
     kwargs={
-        "env_cfg_entry_point": f"{task_entry}.franka_cabinet_env:FrankaCabinetEnvCfg",
+        "env_cfg_entry_point": f"{__name__}.franka_cabinet_env:FrankaCabinetEnvCfg",
         "rl_games_cfg_entry_point": f"{agents.__name__}:rl_games_ppo_cfg.yaml",
         "rsl_rl_cfg_entry_point": f"{agents.__name__}.rsl_rl_ppo_cfg:FrankaCabinetPPORunnerCfg",
         "skrl_cfg_entry_point": f"{agents.__name__}:skrl_ppo_cfg.yaml",
