+++ source/extensions/omni.isaac.lab_tasks/omni/isaac/lab_tasks/direct/humanoid/__init__.py
@@ -15,14 +15,12 @@ from . import agents
 # Register Gym environments.
 ##
 
-task_entry = "omni.isaac.lab_tasks.direct.humanoid"
-
 gym.register(
     id="Isaac-Humanoid-Direct-v0",
-    entry_point=f"{task_entry}.humanoid_env:HumanoidEnv",
+    entry_point=f"{__name__}.humanoid_env:HumanoidEnv",
     disable_env_checker=True,
     kwargs={
-        "env_cfg_entry_point": f"{task_entry}.humanoid_env:HumanoidEnvCfg",
+        "env_cfg_entry_point": f"{__name__}.humanoid_env:HumanoidEnvCfg",
         "rl_games_cfg_entry_point": f"{agents.__name__}:rl_games_ppo_cfg.yaml",
         "rsl_rl_cfg_entry_point": f"{agents.__name__}.rsl_rl_ppo_cfg:HumanoidPPORunnerCfg",
         "skrl_cfg_entry_point": f"{agents.__name__}:skrl_ppo_cfg.yaml",
