+++ source/extensions/omni.isaac.lab_tasks/omni/isaac/lab_tasks/direct/quadcopter/__init__.py
@@ -15,14 +15,12 @@ from . import agents
 # Register Gym environments.
 ##
 
-task_entry = "omni.isaac.lab_tasks.direct.quadcopter"
-
 gym.register(
     id="Isaac-Quadcopter-Direct-v0",
-    entry_point=f"{task_entry}.quadcopter_env:QuadcopterEnv",
+    entry_point=f"{__name__}.quadcopter_env:QuadcopterEnv",
     disable_env_checker=True,
     kwargs={
-        "env_cfg_entry_point": f"{task_entry}.quadcopter_env:QuadcopterEnvCfg",
+        "env_cfg_entry_point": f"{__name__}.quadcopter_env:QuadcopterEnvCfg",
         "rl_games_cfg_entry_point": f"{agents.__name__}:rl_games_ppo_cfg.yaml",
         "rsl_rl_cfg_entry_point": f"{agents.__name__}.rsl_rl_ppo_cfg:QuadcopterPPORunnerCfg",
         "skrl_cfg_entry_point": f"{agents.__name__}:skrl_ppo_cfg.yaml",
