+++ source/extensions/omni.isaac.lab_tasks/omni/isaac/lab_tasks/direct/shadow_hand/__init__.py
@@ -15,7 +15,6 @@ from . import agents
 # Register Gym environments.
 ##
 
-task_entry = "omni.isaac.lab_tasks.direct.shadow_hand"
 inhand_task_entry = "omni.isaac.lab_tasks.direct.inhand_manipulation"
 
 gym.register(
@@ -23,7 +22,7 @@ gym.register(
     entry_point=f"{inhand_task_entry}.inhand_manipulation_env:InHandManipulationEnv",
     disable_env_checker=True,
     kwargs={
-        "env_cfg_entry_point": f"{task_entry}.shadow_hand_env_cfg:ShadowHandEnvCfg",
+        "env_cfg_entry_point": f"{__name__}.shadow_hand_env_cfg:ShadowHandEnvCfg",
         "rl_games_cfg_entry_point": f"{agents.__name__}:rl_games_ppo_cfg.yaml",
         "rsl_rl_cfg_entry_point": f"{agents.__name__}.rsl_rl_ppo_cfg:ShadowHandPPORunnerCfg",
         "skrl_cfg_entry_point": f"{agents.__name__}:skrl_ppo_cfg.yaml",
@@ -35,7 +34,7 @@ gym.register(
     entry_point=f"{inhand_task_entry}.inhand_manipulation_env:InHandManipulationEnv",
     disable_env_checker=True,
     kwargs={
-        "env_cfg_entry_point": f"{task_entry}.shadow_hand_env_cfg:ShadowHandOpenAIEnvCfg",
+        "env_cfg_entry_point": f"{__name__}.shadow_hand_env_cfg:ShadowHandOpenAIEnvCfg",
         "rl_games_cfg_entry_point": f"{agents.__name__}:rl_games_ppo_ff_cfg.yaml",
         "rsl_rl_cfg_entry_point": f"{agents.__name__}.rsl_rl_ppo_cfg:ShadowHandAsymFFPPORunnerCfg",
         "skrl_cfg_entry_point": f"{agents.__name__}:skrl_ff_ppo_cfg.yaml",
@@ -47,7 +46,7 @@ gym.register(
     entry_point=f"{inhand_task_entry}.inhand_manipulation_env:InHandManipulationEnv",
     disable_env_checker=True,
     kwargs={
-        "env_cfg_entry_point": f"{task_entry}.shadow_hand_env_cfg:ShadowHandOpenAIEnvCfg",
+        "env_cfg_entry_point": f"{__name__}.shadow_hand_env_cfg:ShadowHandOpenAIEnvCfg",
         "rl_games_cfg_entry_point": f"{agents.__name__}:rl_games_ppo_lstm_cfg.yaml",
     },
 )
@@ -56,10 +55,10 @@ gym.register(
 
 gym.register(
     id="Isaac-Repose-Cube-Shadow-Vision-Direct-v0",
-    entry_point=f"{task_entry}.shadow_hand_vision_env:ShadowHandVisionEnv",
+    entry_point=f"{__name__}.shadow_hand_vision_env:ShadowHandVisionEnv",
     disable_env_checker=True,
     kwargs={
-        "env_cfg_entry_point": f"{task_entry}.shadow_hand_vision_env:ShadowHandVisionEnvCfg",
+        "env_cfg_entry_point": f"{__name__}.shadow_hand_vision_env:ShadowHandVisionEnvCfg",
         "rsl_rl_cfg_entry_point": f"{agents.__name__}.rsl_rl_ppo_cfg:ShadowHandVisionFFPPORunnerCfg",
         "rl_games_cfg_entry_point": f"{agents.__name__}:rl_games_ppo_vision_cfg.yaml",
     },
@@ -67,10 +66,10 @@ gym.register(
 
 gym.register(
     id="Isaac-Repose-Cube-Shadow-Vision-Direct-Play-v0",
-    entry_point=f"{task_entry}.shadow_hand_vision_env:ShadowHandVisionEnv",
+    entry_point=f"{__name__}.shadow_hand_vision_env:ShadowHandVisionEnv",
     disable_env_checker=True,
     kwargs={
-        "env_cfg_entry_point": f"{task_entry}.shadow_hand_vision_env:ShadowHandVisionEnvPlayCfg",
+        "env_cfg_entry_point": f"{__name__}.shadow_hand_vision_env:ShadowHandVisionEnvPlayCfg",
         "rsl_rl_cfg_entry_point": f"{agents.__name__}.rsl_rl_ppo_cfg:ShadowHandVisionFFPPORunnerCfg",
         "rl_games_cfg_entry_point": f"{agents.__name__}:rl_games_ppo_vision_cfg.yaml",
     },
