+++ source/extensions/omni.isaac.lab_tasks/omni/isaac/lab_tasks/direct/shadow_hand_over/__init__.py
@@ -15,14 +15,12 @@ from . import agents
 # Register Gym environments.
 ##
 
-task_entry = "omni.isaac.lab_tasks.direct.shadow_hand_over"
-
 gym.register(
     id="Isaac-Shadow-Hand-Over-Direct-v0",
-    entry_point=f"{task_entry}.shadow_hand_over_env:ShadowHandOverEnv",
+    entry_point=f"{__name__}.shadow_hand_over_env:ShadowHandOverEnv",
     disable_env_checker=True,
     kwargs={
-        "env_cfg_entry_point": f"{task_entry}.shadow_hand_over_env_cfg:ShadowHandOverEnvCfg",
+        "env_cfg_entry_point": f"{__name__}.shadow_hand_over_env_cfg:ShadowHandOverEnvCfg",
         "rl_games_cfg_entry_point": f"{agents.__name__}:rl_games_ppo_cfg.yaml",
         "skrl_cfg_entry_point": f"{agents.__name__}:skrl_ppo_cfg.yaml",
         "skrl_ippo_cfg_entry_point": f"{agents.__name__}:skrl_ippo_cfg.yaml",
